refactor: Move abc.py and storage system modules to volumes package
