+++ src/ai/backend/storage/volumes/cephfs/__init__.py
@@ -8,10 +8,16 @@ import aiofiles.os
 
 from ai.backend.common.types import BinarySize, QuotaScopeID
 from ai.backend.storage.exception import QuotaScopeNotFoundError
+from ai.backend.storage.subproc import run
+from ai.backend.storage.types import CapacityUsage, Optional, QuotaConfig, QuotaUsage, TreeUsage
 
-from ..abc import CAP_FAST_SIZE, CAP_QUOTA, CAP_VFOLDER, AbstractFSOpModel, AbstractQuotaModel
-from ..subproc import run
-from ..types import CapacityUsage, Optional, QuotaConfig, QuotaUsage, TreeUsage
+from ..abc import (
+    CAP_FAST_SIZE,
+    CAP_QUOTA,
+    CAP_VFOLDER,
+    AbstractFSOpModel,
+    AbstractQuotaModel,
+)
 from ..vfs import BaseFSOpModel, BaseQuotaModel, BaseVolume
