+++ src/ai/backend/storage/volumes/purestorage/__init__.py
@@ -8,9 +8,15 @@ from typing import FrozenSet
 
 from ai.backend.common.types import HardwareMetadata
 from ai.backend.logging import BraceStyleAdapter
+from ai.backend.storage.types import CapacityUsage, FSPerfMetric
 
-from ..abc import CAP_FAST_FS_SIZE, CAP_FAST_SCAN, CAP_METRIC, CAP_VFOLDER, AbstractFSOpModel
-from ..types import CapacityUsage, FSPerfMetric
+from ..abc import (
+    CAP_FAST_FS_SIZE,
+    CAP_FAST_SCAN,
+    CAP_METRIC,
+    CAP_VFOLDER,
+    AbstractFSOpModel,
+)
 from ..vfs import BaseVolume
 from .purity import PurityClient
 from .rapidfiles import RapidFileToolsFSOpModel
