+++ src/ai/backend/storage/volumes/purestorage/rapidfiles.py
@@ -6,10 +6,10 @@ from subprocess import CalledProcessError
 from typing import AsyncIterator
 
 from ai.backend.common.types import BinarySize
+from ai.backend.storage.subproc import run
+from ai.backend.storage.types import DirEntry, DirEntryType, Stat, TreeUsage
+from ai.backend.storage.utils import fstime2datetime
 
-from ..subproc import run
-from ..types import DirEntry, DirEntryType, Stat, TreeUsage
-from ..utils import fstime2datetime
 from ..vfs import BaseFSOpModel
